feat: more refactorings & rename test.py to group.py
