+++ src/group.py
@@ -1,0 +1,230 @@
+from datetime import timedelta
+
+import boto3
+from mypy_boto3_identitystore import IdentityStoreClient
+from mypy_boto3_sso_admin import SSOAdminClient
+from slack_bolt import Ack, BoltContext
+from slack_sdk import WebClient
+
+from slack_sdk.web.slack_response import SlackResponse
+
+import access_control
+import config
+import entities
+import schedule
+import slack_helpers
+import sso
+from errors import handle_errors
+
+logger = config.get_logger(service="main")
+cfg = config.get_config()
+
+
+identity_store_client: IdentityStoreClient = boto3.client("identitystore", region_name="us-east-1")
+sso_client: SSOAdminClient = boto3.client("sso-admin", region_name="us-east-1")
+schedule_client = boto3.client("scheduler", region_name="us-east-1")
+sso_instance = sso.describe_sso_instance(sso_client, cfg.sso_instance_arn)
+identity_store_id = sso_instance.identity_store_id
+
+
+@handle_errors
+def handle_request_for_group_access_submittion(
+    body: dict,
+    ack: Ack,  # noqa: ARG001
+    client: WebClient,
+    context: BoltContext,  # noqa: ARG001
+) -> SlackResponse | None:
+    logger.info("Handling request for access submittion")
+    request = slack_helpers.RequestForGroupAccessView.parse(body)
+    logger.info("View submitted", extra={"view": request})
+    requester = slack_helpers.get_user(client, id=request.requester_slack_id)
+
+    group = sso.describe_group(identity_store_id, request.group_id, identity_store_client)
+
+    decision = access_control.make_decision_on_access_request(
+        cfg.group_statements,
+        requester_email=requester.email,
+        group_id=request.group_id,
+    )
+
+    show_buttons = bool(decision.approvers)
+    slack_response = client.chat_postMessage(
+        blocks=slack_helpers.build_approval_request_message_blocks(
+            requester_slack_id=request.requester_slack_id,
+            group=group,
+            reason=request.reason,
+            permission_duration=request.permission_duration,
+            show_buttons=show_buttons,
+            color_coding_emoji=cfg.waiting_result_emoji,
+        ),
+        channel=cfg.slack_channel_id,
+        text=f"Request for access to {group.name} group from {requester.real_name}",
+    )
+
+    if show_buttons:
+        ts = slack_response["ts"]
+        if ts is not None:
+            schedule.schedule_discard_buttons_event(
+                schedule_client=schedule_client, #type: ignore # noqa: PGH003
+                time_stamp=ts,
+                channel_id=cfg.slack_channel_id,
+            )
+            schedule.schedule_approver_notification_event(
+                schedule_client=schedule_client, #type: ignore # noqa: PGH003
+                message_ts=ts,
+                channel_id=cfg.slack_channel_id,
+                time_to_wait=timedelta(
+                    minutes=cfg.approver_renotification_initial_wait_time,
+                ),
+            )
+
+    match decision.reason:
+        case access_control.DecisionReason.ApprovalNotRequired:
+            text = "Approval for this Group is not required. Request will be approved automatically."
+            color_coding_emoji = cfg.good_result_emoji
+        case access_control.DecisionReason.SelfApproval:
+            text = "Self approval is allowed and requester is an approver. Request will be approved automatically."
+            color_coding_emoji = cfg.good_result_emoji
+        case access_control.DecisionReason.RequiresApproval:
+            approvers = [slack_helpers.get_user_by_email(client, email) for email in decision.approvers]
+            mention_approvers = " ".join(f"<@{approver.id}>" for approver in approvers)
+            text = f"{mention_approvers} there is a request waiting for the approval."
+            color_coding_emoji = cfg.waiting_result_emoji
+        case access_control.DecisionReason.NoApprovers:
+            text = "Nobody can approve this request."
+            color_coding_emoji = cfg.bad_result_emoji
+        case access_control.DecisionReason.NoStatements:
+            text = "There are no statements for this Group."
+            color_coding_emoji = cfg.bad_result_emoji
+
+    client.chat_postMessage(text=text, thread_ts=slack_response["ts"], channel=cfg.slack_channel_id)
+
+    blocks = slack_helpers.HeaderSectionBlock.set_color_coding(
+        blocks=slack_response["message"]["blocks"],
+        color_coding_emoji=color_coding_emoji,
+    )
+    client.chat_update(
+        channel=cfg.slack_channel_id,
+        ts=slack_response["ts"],
+        blocks=blocks,
+        text=text,
+    )
+
+    user_principal_id = sso.get_user_principal_id_by_email(identity_store_client, sso_instance.identity_store_id, requester.email)
+
+    access_control.execute_decision_on_group_request(
+        group = group,
+        user_principal_id = user_principal_id,
+        permission_duration = request.permission_duration,
+        approver = requester,
+        requester = requester,
+        reason = request.reason,
+        decision = decision,
+        identity_store_id=identity_store_id,
+    )
+
+    if decision.grant:
+
+        client.chat_postMessage(
+            channel=cfg.slack_channel_id,
+            text=f"Permissions granted to <@{requester.id}>",
+            thread_ts=slack_response["ts"],
+        )
+
+cache_for_dublicate_requests = {}
+
+
+@handle_errors
+def handle_group_button_click(body: dict, client: WebClient, context: BoltContext) -> SlackResponse:  #type: ignore # noqa: PGH003 ARG001
+    logger.info("Handling button click")
+    payload = slack_helpers.ButtonGroupClickedPayload.parse_obj(body)
+    logger.info("Button click payload", extra={"payload": payload})
+    approver = slack_helpers.get_user(client, id=payload.approver_slack_id)
+    requester = slack_helpers.get_user(client, id=payload.request.requester_slack_id)
+
+    if (
+        cache_for_dublicate_requests.get("requester_slack_id") == payload.request.requester_slack_id
+        and cache_for_dublicate_requests["group_id"] == payload.request.group_id
+    ):
+        return client.chat_postMessage(
+            channel=payload.channel_id,
+            text=f"<@{approver.id}> request is already in progress, please wait for the result.",
+            thread_ts=payload.thread_ts,
+        )
+    cache_for_dublicate_requests["requester_slack_id"] = payload.request.requester_slack_id
+    cache_for_dublicate_requests["group_id"] = payload.request.group_id
+
+
+    if payload.action == entities.ApproverAction.Discard:
+        blocks = slack_helpers.HeaderSectionBlock.set_color_coding(
+            blocks=payload.message["blocks"],
+            color_coding_emoji=cfg.bad_result_emoji,
+        )
+
+        blocks = slack_helpers.remove_blocks(blocks, block_ids=["buttons"])
+        blocks.append(slack_helpers.button_click_info_block(payload.action, approver.id).to_dict())
+
+        text = f"Request was discarded by<@{approver.id}> "
+        client.chat_update(
+            channel=payload.channel_id,
+            ts=payload.thread_ts,
+            blocks=blocks,
+            text=text,
+        )
+
+        cache_for_dublicate_requests.clear()
+        return client.chat_postMessage(
+            channel=payload.channel_id,
+            text=text,
+            thread_ts=payload.thread_ts,
+        )
+
+    decision = access_control.make_decision_on_approve_request(
+        action=payload.action,
+        statements=cfg.group_statements, #type: ignore # noqa: PGH003
+        group_id=payload.request.group_id,
+        approver_email=approver.email,
+        requester_email=requester.email,
+    )
+
+    logger.info("Decision on request was made", extra={"decision": decision})
+
+    if not decision.permit:
+        cache_for_dublicate_requests.clear()
+        return client.chat_postMessage(
+            channel=payload.channel_id,
+            text=f"<@{approver.id}> you can not approve this request",
+            thread_ts=payload.thread_ts,
+        )
+
+    text = f"Permissions granted to <@{requester.id}> by <@{approver.id}>."
+    blocks = slack_helpers.HeaderSectionBlock.set_color_coding(
+        blocks=payload.message["blocks"],
+        color_coding_emoji=cfg.good_result_emoji,
+    )
+
+    blocks = slack_helpers.remove_blocks(blocks, block_ids=["buttons"])
+    blocks.append(slack_helpers.button_click_info_block(payload.action, approver.id).to_dict())
+    client.chat_update(
+        channel=payload.channel_id,
+        ts=payload.thread_ts,
+        blocks=blocks,
+        text=text,
+    )
+
+    access_control.execute_decision_on_group_request(
+        decision=decision,
+        group = sso.describe_group(identity_store_id, payload.request.group_id, identity_store_client),
+        user_principal_id = sso.get_user_principal_id_by_email(identity_store_client, sso_instance.identity_store_id, requester.email),
+        permission_duration=payload.request.permission_duration,
+        approver=approver,
+        requester=requester,
+        reason=payload.request.reason,
+        identity_store_id=identity_store_id
+    )
+    cache_for_dublicate_requests.clear()
+    return client.chat_postMessage(
+        channel=payload.channel_id,
+        text=text,
+        thread_ts=payload.thread_ts,
+    )
